fixup: Move TxIssueStatus as per review comment
